refactored code to avoid code duplication in annotation moving
