+++ src/neuroglancer/display_context.ts
@@ -19,6 +19,7 @@ import {RefCounted} from 'neuroglancer/util/disposable';
 import {NullarySignal} from 'neuroglancer/util/signal';
 import {WatchableVisibilityPriority} from 'neuroglancer/visibility_priority/frontend';
 import {GL, initializeWebGL} from 'neuroglancer/webgl/context';
+import {vec3} from 'neuroglancer/util/geom';
 
 export abstract class RenderedPanel extends RefCounted {
   gl: GL;
@@ -59,7 +60,9 @@ export abstract class RenderedPanel extends RefCounted {
   abstract onResize(): void;
 
   abstract draw(): void;
-
+  
+  abstract translateDataPointByViewportPixels(out: vec3, orig: vec3, deltaX: number, deltaY: number): vec3;
+  
   disposed() {
     this.context.removePanel(this);
     super.disposed();
